created BoardElement & EndCard classes to manage individual html elements & reduce complexity of WebChessGame
