+++ docs/index.html
@@ -10,12 +10,8 @@
 <body>
 
     <div class="main-container">
-
         <div class="board-container">
-
         </div>
-
-
     </div>
