+++ src/main.ts
@@ -2,7 +2,8 @@
 import { WebChessGame } from './webchessgame.js';
 
 const boardContainer: HTMLElement | null = document.querySelector(".board-container")
+const mainContainer: HTMLElement | null = document.querySelector(".main-container")
 
-if (boardContainer){
-    new WebChessGame(boardContainer);
+if (mainContainer && boardContainer){
+    new WebChessGame(mainContainer, boardContainer);
 }
